+++ pkg/boundedwaitgroup/boundedwaitgroup_test.go
@@ -1,4 +1,4 @@
-package util
+package boundedwaitgroup
 
 import (
 	"sync"
