+++ tempodb/retention.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/go-kit/kit/log/level"
 
-	"github.com/grafana/tempo/pkg/util"
+	"github.com/grafana/tempo/pkg/boundedwaitgroup"
 )
 
 // todo: pass a context/chan in to cancel this cleanly
@@ -20,7 +20,7 @@ func (rw *readerWriter) retentionLoop() {
 func (rw *readerWriter) doRetention() {
 	tenants := rw.blocklistTenants()
 
-	bg := util.NewBoundedWaitGroup(rw.compactorCfg.RetentionConcurrency)
+	bg := boundedwaitgroup.NewBoundedWaitGroup(rw.compactorCfg.RetentionConcurrency)
 
 	for _, tenantID := range tenants {
 		bg.Add(1)
